+++ docs/src/markdown/about/changelog.md
@@ -2,6 +2,9 @@
 
 ## 1.9.4
 
+- **FIX**: `:checked` does not require an `option` element to be a child of a `select` element for it to match.
+- **FIX**: `:disabled` should only select `option` elements under a disabled `optgroup` if that `optgroup` is the
+closest parent.
 - **FIX**: Fix `:lang()` wildcard matching when applied to the primary language tag (i.e. `:lang('*-US')`). When a
 wildcard is applied at the start of the pattern, it should match zero or more, not one or more.